elgen-esg-online-model-service/requirements.txt
requests~=2.31.0
loguru==0.7.0
pydantic~=1.10.9
dependency-injector~=4.41.0
fastapi~=0.96.0
uvicorn~=0.22.0
opentelemetry-instrumentation-logging==0.38b0
opentelemetry-exporter-otlp-proto-grpc==1.18.0
opentelemetry-sdk==1.18.0
python-json-logger==2.0.7
pytest~=7.4.0
pytest-cov~=4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.11.1
requests~=2.31.0
botocore~=1.29.101
httpx~=0.24.1